I think this is about as bomb-proof as it can get...

+++ src/piwiz.c
@@ -66,6 +66,9 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define SKIP_BTN 1
 #define PREV_BTN 2
 
+#define FORWARD 1
+#define BACKWARD -1
+
 // columns in localisation list stores
 
 #define CL_CNAME 0
@@ -394,7 +397,8 @@ static gboolean clock_synced (void);
 static void resync (void);
 static gboolean ntp_check (gpointer data);
 static void page_changed (GtkNotebook *notebook, GtkWidget *page, int pagenum, gpointer data);
-static void page_forward (void);
+static gboolean page_shown (int page);
+static void change_page (int dir);
 static void next_page (GtkButton* btn, gpointer ptr);
 static void prev_page (GtkButton* btn, gpointer ptr);
 static void skip_page (GtkButton* btn, gpointer ptr);
@@ -1124,7 +1128,7 @@ void connect_success (void)
         g_source_remove (conn_timeout);
         conn_timeout = 0;
         hide_message ();
-        page_forward ();
+        change_page (FORWARD);
     }
 }
 
@@ -2021,56 +2025,42 @@ static void page_changed (GtkNotebook *notebook, GtkWidget *page, int pagenum, g
     }
 }
 
-static void page_forward (void)
+static gboolean page_shown (int page)
 {
-    int next;
-
-    switch (gtk_notebook_get_current_page (GTK_NOTEBOOK (wizard_nb)))
+    switch (page)
     {
-        case PAGE_INTRO :   next = PAGE_LOCALE;
-                            break;
-
-        case PAGE_LOCALE :  next = PAGE_PASSWD;
-                            break;
-
-        case PAGE_PASSWD :  if (chuser != NULL)
-                            {
-                                next = PAGE_DONE;
-                                break;
-                            }
-                            else if (is_pi && wm == WM_OPENBOX)
-                            {
-                                next = PAGE_OSCAN;
-                                break;
-                            }
-
-        case PAGE_OSCAN :   if (wifi_if)
-                            {
-                                next = PAGE_WIFIAP;
-                                break;
-                            }
+        case PAGE_OSCAN :   return (is_pi && wm == WM_OPENBOX);
 
         case PAGE_WIFIAP :
-        case PAGE_WIFIPSK : if (browser)
-                            {
-                                next = PAGE_BROWSER;
-                                break;
-                            }
+        case PAGE_WIFIPSK : if (wifi_if) return TRUE;
+                            else return FALSE;
 
-        case PAGE_BROWSER : if (rpc)
-                            {
-                                next = PAGE_RPC;
-                                break;
-                            }
+        case PAGE_BROWSER : return browser;
 
-        case PAGE_RPC :     next = PAGE_UPDATE;
-                            break;
+        case PAGE_RPC :     return rpc;
 
+        case PAGE_INTRO :
+        case PAGE_LOCALE :
+        case PAGE_PASSWD :
         case PAGE_UPDATE :
-        default :           next = PAGE_DONE;
-                            break;
+        case PAGE_DONE :
+        default :           return TRUE;
     }
-    gtk_notebook_set_current_page (GTK_NOTEBOOK (wizard_nb), next);
+}
+
+static void change_page (int dir)
+{
+    int page = gtk_notebook_get_current_page (GTK_NOTEBOOK (wizard_nb)) + dir;
+
+    for (; dir == -1 ? page >= PAGE_INTRO : page < PAGE_DONE; page += dir)
+    {
+        if (page == PAGE_WIFIPSK) page += dir;  // only jump to this one explicitly when needed
+        if (page_shown (page))
+        {
+            gtk_notebook_set_current_page (GTK_NOTEBOOK (wizard_nb), page);
+            break;
+        }
+    }
 }
 
 static void next_page (GtkButton* btn, gpointer ptr)
@@ -2090,7 +2080,7 @@ static void next_page (GtkButton* btn, gpointer ptr)
                                 if (wm != WM_OPENBOX) vsystem ("wfpanelctl bluetooth apstop");
                                 else vsystem ("lxpanelctl command bluetooth apstop");
                             }
-                            page_forward ();
+                            change_page (FORWARD);
                             break;
 
         case PAGE_LOCALE :  // get the combo entries and look up relevant codes in database
@@ -2137,7 +2127,7 @@ static void next_page (GtkButton* btn, gpointer ptr)
                                 message (_("Setting location - please wait..."), 0, 0, -1, TRUE);
                                 g_thread_new (NULL, set_locale, NULL);
                             }
-                            else page_forward ();
+                            else change_page (FORWARD);
                             break;
 
         case PAGE_PASSWD :  if (user != NULL)
@@ -2201,18 +2191,19 @@ static void next_page (GtkButton* btn, gpointer ptr)
                             }
                             user = g_strdup (gtk_entry_get_text (GTK_ENTRY (user_te)));
                             pw = g_strdup (crypt (gtk_entry_get_text (GTK_ENTRY (pwd1_te)), crypt_gensalt (NULL, 0, NULL, 0)));
-                            page_forward ();
+                            if (chuser != NULL) gtk_notebook_set_current_page (GTK_NOTEBOOK (wizard_nb), PAGE_DONE);
+                            else change_page (FORWARD);
                             break;
 
-        case PAGE_OSCAN :   page_forward ();
+        case PAGE_OSCAN :   change_page (FORWARD);
                             break;
 
         case PAGE_WIFIAP :  if (ssid) g_free (ssid);
                             ssid = NULL;
-                            if (!find_line (&ssid, &secure, &connected)) page_forward ();
+                            if (!find_line (&ssid, &secure, &connected)) change_page (FORWARD);
                             else
                             {
-                                if (connected) page_forward ();
+                                if (connected) change_page (FORWARD);
                                 else if (secure)
                                 {
                                     text = g_strdup_printf (_("Enter the password for the WiFi network \"%s\"."), ssid);
@@ -2242,14 +2233,14 @@ static void next_page (GtkButton* btn, gpointer ptr)
                                 vsystem ("sudo raspi-config nonint do_browser chromium-browser pi");
                             else
                                 vsystem ("sudo raspi-config nonint do_browser firefox pi");
-                            page_forward ();
+                            change_page (FORWARD);
                             break;
 
         case PAGE_RPC :     if (gtk_switch_get_active (GTK_SWITCH (rpc_sw)))
                                 vsystem ("sudo systemctl --global enable rpi-connect;sudo systemctl --global enable rpi-connect-wayvnc");
                             else
                                 vsystem ("sudo systemctl --global disable rpi-connect;sudo systemctl --global disable rpi-connect-wayvnc");
-                            page_forward ();
+                            change_page (FORWARD);
                             break;
 
         case PAGE_UPDATE :  if (net_available ())
@@ -2279,60 +2270,21 @@ static void next_page (GtkButton* btn, gpointer ptr)
 
 static void prev_page (GtkButton* btn, gpointer ptr)
 {
-    int prev;
-
     last_btn = PREV_BTN;
-    switch (gtk_notebook_get_current_page (GTK_NOTEBOOK (wizard_nb)))
+
+    if (chuser != NULL)
     {
-        case PAGE_DONE :    if (chuser != NULL)
-                                prev = PAGE_PASSWD;
-                            else
-                                prev = PAGE_UPDATE;
-                            break;
-
-        case PAGE_UPDATE :  if (rpc)
-                            {
-                                prev = PAGE_RPC;
-                                break;
-                            }
-
-        case PAGE_RPC :     if (browser)
-                            {
-                                prev = PAGE_BROWSER;
-                                break;
-                            }
-
-        case PAGE_BROWSER :
-        case PAGE_WIFIPSK : if (wifi_if)
-                            {
-                                prev = PAGE_WIFIAP;
-                                break;
-                            }
-
-        case PAGE_WIFIAP :  if (is_pi && wm  == WM_OPENBOX)
-                            {
-                                prev = PAGE_OSCAN;
-                                break;
-                            }
-
-        case PAGE_OSCAN :   prev = PAGE_PASSWD;
-                            break;
-
-        case PAGE_PASSWD :  if (chuser != NULL)
-                            {
-                                // call the script to cancel renaming and reset original user
-                                vsystem ("sudo /usr/bin/cancel-rename %s", chuser);
-                                vsystem ("sync;reboot");
-                                gtk_main_quit ();
-                            }
-                            else prev = PAGE_LOCALE;
-                            break;
-
-        case PAGE_LOCALE :
-        default :           prev = PAGE_INTRO;
-                            break;
+        int page = gtk_notebook_get_current_page (GTK_NOTEBOOK (wizard_nb));
+        if (page == PAGE_DONE) gtk_notebook_set_current_page (GTK_NOTEBOOK (wizard_nb), PAGE_PASSWD);
+        else if (page == PAGE_PASSWD)
+        {
+            // call the script to cancel renaming and reset original user
+            vsystem ("sudo /usr/bin/cancel-rename %s", chuser);
+            vsystem ("sync;reboot");
+            gtk_main_quit ();
+        }
     }
-    gtk_notebook_set_current_page (GTK_NOTEBOOK (wizard_nb), prev);
+    else change_page (BACKWARD);
 }
 
 static void skip_page (GtkButton* btn, gpointer ptr)
@@ -2341,7 +2293,7 @@ static void skip_page (GtkButton* btn, gpointer ptr)
     switch (gtk_notebook_get_current_page (GTK_NOTEBOOK (wizard_nb)))
     {
         case PAGE_WIFIAP :
-        case PAGE_WIFIPSK : page_forward ();
+        case PAGE_WIFIPSK : change_page (FORWARD);
                             break;
 
         case PAGE_UPDATE :  gchar *buf = g_strdup_printf ("check-language-support -l %s_%s", lc, cc);
@@ -2353,7 +2305,7 @@ static void skip_page (GtkButton* btn, gpointer ptr)
                                 message (_("If installing updates is skipped, translation files will not be installed."), 1, PAGE_DONE, -1, FALSE);
                             else if (uninst)
                                 message (_("If installing updates is skipped, the unused browser will not be uninstalled."), 1, PAGE_DONE, -1, FALSE);
-                            else page_forward ();
+                            else change_page (FORWARD);
                             g_free (buf);
                             g_free (lpack);
                             break;
